+++ EchoLedger/Features/Transaction/Domain/UseCases/UpdateTransactionInput.swift
@@ -13,7 +13,7 @@ struct UpdateTransactionInput {
     let userId: UUID
     let label: String
     let date: Date
-    let totalAmount: Decimal
+    let totalAmount: Double
     let note: String?
     let isExpense: Bool
     let category: TransactionCategory
